test: quebra o arquivo por assunto e nomeia o que a guarda do Jsonb não fecha
O arquivo de teste tinha 571 linhas e quebrava o portão do próprio repo
(`test_max_lines_python.py`, 350 linhas, §0.5). Virou três por assunto —
topo não-objeto, campo filho, valor não finito — mais um módulo de fixtures
`_`-prefixado, no padrão que `tests/_billing_grants_helpers.py` já usa.

Não virou `conftest.py` de propósito: `configured_admin` é autouse, e num
conftest passaria a trocar a senha do admin e a resetar o `limiter._storage`
da suíte inteira.

Node IDs coletados antes e depois da quebra: idênticos, com os parâmetros —
nenhum dos 97 sumiu. Um controle negativo por arquivo (15/8/3 vermelhos),
porque o risco real de quebrar um arquivo de teste é um caso ficar verde de
graça por fixture perdida.

E uma cláusula no comentário do webhook: a recusa no parse fecha a metade
NUMÉRICA da via do `Jsonb`. NUL e surrogate solitário, no valor ou na chave,
seguem chegando ao `Jsonb(raw)`, ao param `text` do item_id e à lista do
`any(%s)` — 500, pré-existente, só com o secret. Fica nomeado para o próximo
leitor não concluir que a via está fechada.
